+++ CHANGELOG.md
@@ -33,7 +33,7 @@ If you love `detect-secrets`, please star our project on GitHub to show your sup
 
 #### :tada: New Features
 
-- Add support to extract pragmas in `scan_diff` which is called by [`detect-secrets-server`](https://github.com/Yelp/detect-secrets-server) (by [@francisluz]) 
+- Add support to extract pragmas in `scan_diff` which is called by [`detect-secrets-server`](https://github.com/Yelp/detect-secrets-server) (by [@francisluz])
 
 ##### August 27th, 2020